Setelah refactor data clumps v2 mengakibatkan class employee jadi long class, maka class identity dibuat class baru(file baru)
